+++ EnterTheMUngeons/finalBoss.c
@@ -22,7 +22,7 @@ void bossFinalCreation() {
 
 void finalBoss() {
 	final.counter[2]++;
-	srand(time(0));
+	srand((unsigned int)time(0));
 	if (final.counter[2] == 180) {
 		final.counter[2] = 0;
 		int x = rand() % 2;
@@ -77,10 +77,10 @@ void finalBoss() {
 		boss.bossRect[4].y = WINDOW_HEIGHT / 2 - 50;
 	}
 	finalAtakeakSortu(playerRect.player, playerRect.Bala);
-	ValorRects(&boss.bossLifeFrame, boss.bossRect[4].w, 20, boss.bossRect[4].x, boss.bossRect[4].y - 20);
-	ValorRects(&boss.bossLife, final.health[0], 9, boss.bossRect[4].x + 16, boss.bossRect[4].y - 14);
-	ValorRects(&boss.bossLife2, final.health[1], 9, boss.bossRect[4].x + 16, boss.bossRect[4].y - 14);
-	ValorRects(&boss.bossLife3, final.health[2], 9, boss.bossRect[4].x + 16, boss.bossRect[4].y - 14);
+	ValorRects(&boss.bossLifeFrame,(int)boss.bossRect[4].w, 20, boss.bossRect[4].x, boss.bossRect[4].y - 20);
+	ValorRects(&boss.bossLife, (int)final.health[0], 9, boss.bossRect[4].x + 16, boss.bossRect[4].y - 14);
+	ValorRects(&boss.bossLife2, (int)final.health[1], 9, boss.bossRect[4].x + 16, boss.bossRect[4].y - 14);
+	ValorRects(&boss.bossLife3, (int)final.health[2], 9, boss.bossRect[4].x + 16, boss.bossRect[4].y - 14);
 	enemySpawn.enemykilled[0] = 0;
 	
 }
@@ -95,9 +95,9 @@ void finalAtakeakSortu(SDL_Rect player, SDL_Rect bala) {
 			final.finalVelY[0] = ATAQUEVEL - 4;
 			final.finalVelX[0] = ATAQUEVEL - 4;
 			final.finalHasBeenCreated[0] = 1;
-			final.deltax = ((player).x + (player.w / 2)) - (boss.bossRect[4].x + boss.bossRect[4].w / 2);
-			final.deltay = ((player).y + (player.h / 2)) - (boss.bossRect[4].y + boss.bossRect[4].h / 2);
-			final.distance = sqrt(final.deltax*final.deltax + final.deltay * final.deltay);
+			final.deltax = (float)((player).x + (player.w / 2)) - (boss.bossRect[4].x + boss.bossRect[4].w / 2);
+			final.deltay = (float)((player).y + (player.h / 2)) - (boss.bossRect[4].y + boss.bossRect[4].h / 2);
+			final.distance = (float)sqrt(final.deltax*final.deltax + final.deltay * final.deltay);
 		}
 		if (final.finalHasBeenCreated[1] == 0 && final.counter[1] == 0) {
 			final.finalRect[1].h = 26;
@@ -107,9 +107,9 @@ void finalAtakeakSortu(SDL_Rect player, SDL_Rect bala) {
 			final.finalVelY[1] = ATAQUEVEL - 4;
 			final.finalVelX[1] = ATAQUEVEL - 4;
 			final.finalHasBeenCreated[1] = 1;
-			final.deltax = ((player).x + (player.w / 2)) - (boss.bossRect[4].x + boss.bossRect[4].w / 2);
-			final.deltay = ((player).y + (player.h / 2)) - (boss.bossRect[4].y + boss.bossRect[4].h / 2);
-			final.distance = sqrt(final.deltax*final.deltax + final.deltay * final.deltay);
+			final.deltax = (float)((player).x + (player.w / 2)) - (boss.bossRect[4].x + boss.bossRect[4].w / 2);
+			final.deltay = (float)((player).y + (player.h / 2)) - (boss.bossRect[4].y + boss.bossRect[4].h / 2);
+			final.distance = (float)sqrt(final.deltax*final.deltax + final.deltay * final.deltay);
 		}
 	}
 	finalAtakeakMugitu(player, bala);
@@ -125,8 +125,8 @@ void finalAtakeakMugitu(SDL_Rect dest, SDL_Rect bala) {
 			if (final.finalHasBeenCreated[0] == 1) {
 				final.finalVelX[0] = (final.deltax*(ATAQUEVEL)) / final.distance;
 				final.finalVelY[0] = (final.deltay*(ATAQUEVEL)) / final.distance;
-				final.finalRect[0].y += final.finalVelY[0];
-				final.finalRect[0].x += final.finalVelX[0];
+				final.finalRect[0].y += (int)final.finalVelY[0];
+				final.finalRect[0].x += (int)final.finalVelX[0];
 				if (!lifestruct.invencivilty) {
 					if (collisioncheck(final.finalRect[0], dest)) {
 						final.finalHasBeenCreated[0] = 0;
@@ -142,8 +142,8 @@ void finalAtakeakMugitu(SDL_Rect dest, SDL_Rect bala) {
 			if (final.finalHasBeenCreated[1] == 1) {
 				final.finalVelX[1] = (final.deltax*(ATAQUEVEL)) / final.distance;
 				final.finalVelY[1] = (final.deltay*(ATAQUEVEL)) / final.distance;
-				final.finalRect[1].y += final.finalVelY[1];
-				final.finalRect[1].x += final.finalVelX[1];
+				final.finalRect[1].y += (int)final.finalVelY[1];
+				final.finalRect[1].x += (int)final.finalVelX[1];
 				if (!lifestruct.invencivilty) {
 					if (collisioncheck(final.finalRect[1], dest)) {
 						final.finalHasBeenCreated[1] = 0;
@@ -201,36 +201,36 @@ void finalAtakeakMugitu(SDL_Rect dest, SDL_Rect bala) {
 			
 			final.finalVelX[0] = ATAQUEVEL;
 			final.finalVelY[0] = -ATAQUEVEL;
-			final.finalRect[0].y += final.finalVelY[0];
-			final.finalRect[0].x += final.finalVelX[0];
+			final.finalRect[0].y += (int)final.finalVelY[0];
+			final.finalRect[0].x += (int)final.finalVelX[0];
 			final.finalVelX[1] = ATAQUEVEL;
 			final.finalVelY[1] = 0;
-			final.finalRect[1].y += final.finalVelY[1];
-			final.finalRect[1].x += final.finalVelX[1];
+			final.finalRect[1].y += (int)final.finalVelY[1];
+			final.finalRect[1].x += (int)final.finalVelX[1];
 			final.finalVelX[2] = ATAQUEVEL;
 			final.finalVelY[2] = ATAQUEVEL;
-			final.finalRect[2].y += final.finalVelY[2];
-			final.finalRect[2].x += final.finalVelX[2];
+			final.finalRect[2].y += (int)final.finalVelY[2];
+			final.finalRect[2].x += (int)final.finalVelX[2];
 			final.finalVelX[3] = 0;
 			final.finalVelY[3] = ATAQUEVEL;
-			final.finalRect[3].y += final.finalVelY[3];
-			final.finalRect[3].x += final.finalVelX[3];
+			final.finalRect[3].y += (int)final.finalVelY[3];
+			final.finalRect[3].x += (int)final.finalVelX[3];
 			final.finalVelX[4] = -ATAQUEVEL;
 			final.finalVelY[4] = ATAQUEVEL;
-			final.finalRect[4].y += final.finalVelY[4];
-			final.finalRect[4].x += final.finalVelX[4];
+			final.finalRect[4].y += (int)final.finalVelY[4];
+			final.finalRect[4].x += (int)final.finalVelX[4];
 			final.finalVelX[5] = -ATAQUEVEL;
 			final.finalVelY[5] = 0;
-			final.finalRect[5].y += final.finalVelY[5];
-			final.finalRect[5].x += final.finalVelX[5];
+			final.finalRect[5].y += (int)final.finalVelY[5];
+			final.finalRect[5].x += (int)final.finalVelX[5];
 			final.finalVelX[6] = -ATAQUEVEL;
 			final.finalVelY[6] = -ATAQUEVEL;
-			final.finalRect[6].y += final.finalVelY[6];
-			final.finalRect[6].x += final.finalVelX[6];
+			final.finalRect[6].y += (int)final.finalVelY[6];
+			final.finalRect[6].x += (int)final.finalVelX[6];
 			final.finalVelX[7] = 0;
 			final.finalVelY[7] = -ATAQUEVEL;
-			final.finalRect[7].y += final.finalVelY[7];
-			final.finalRect[7].x += final.finalVelX[7];
+			final.finalRect[7].y += (int)final.finalVelY[7];
+			final.finalRect[7].x += (int)final.finalVelX[7];
 
 			if (!lifestruct.invencivilty) {
 				if (collisioncheck(final.finalRect[0], dest) || collisioncheck(final.finalRect[1], dest) || collisioncheck(final.finalRect[2], dest) || collisioncheck(final.finalRect[3], dest) || collisioncheck(final.finalRect[4], dest) || collisioncheck(final.finalRect[5], dest) || collisioncheck(final.finalRect[6], dest) || collisioncheck(final.finalRect[7], dest)) {
@@ -248,9 +248,9 @@ void finalAtakeakMugitu(SDL_Rect dest, SDL_Rect bala) {
 	if (collisioncheck(bala, boss.bossRect[4])) {
 		boss.healthPoints[4] -= 1;
 		
-		if(boss.healthPoints[4] > 200)final.health[0] -= 1.18;
-		if (boss.healthPoints[4] > 100 && boss.healthPoints[4] <= 200)final.health[1] -= 1.18;
-		if (boss.healthPoints[4] <= 100)final.health[2] -= 1.18;
+		if(boss.healthPoints[4] > 200)final.health[0] -= (float)1.18;
+		if (boss.healthPoints[4] > 100 && boss.healthPoints[4] <= 200)final.health[1] -= (float)1.18;
+		if (boss.healthPoints[4] <= 100)final.health[2] -= (float)1.18;
 		if (boss.healthPoints[4] == 0) {
 			levelmanagerstruct.playing = 3;
 			lifestruct.score += 1000;
